refactor(kolme-runtime): extract runtime interfaces module file (#1964)
