+++ crates/kamn-core/src/kolme_runtime_commit.rs
@@ -112,6 +112,7 @@ mod finality_checker;
 mod fork_finality_resolver;
 mod http_transport;
 mod in_memory_client;
+mod interfaces;
 mod live_provider;
 mod notifications_consumer;
 mod notifications_websocket;
@@ -138,6 +139,12 @@ pub use finality_checker::KolmeRuntimeCommitFinalityChecker;
 pub use fork_finality_resolver::KolmeRuntimeCommitForkFinalityResolver;
 pub use http_transport::KolmeRuntimeCommitHttpTransport;
 pub use in_memory_client::InMemoryKolmeRuntimeCommitClient;
+pub use interfaces::{
+    KolmeRuntimeCommitBlockFallbackTransport, KolmeRuntimeCommitClient,
+    KolmeRuntimeCommitFinalityTransport, KolmeRuntimeCommitNotificationsConnection,
+    KolmeRuntimeCommitNotificationsConnector, KolmeRuntimeCommitProvider,
+    KolmeRuntimeCommitProviderTransport,
+};
 pub use live_provider::KolmeRuntimeCommitLiveProvider;
 pub use notifications_consumer::KolmeRuntimeCommitNotificationsConsumer;
 pub use notifications_websocket::{
@@ -152,79 +159,6 @@ pub use request_envelope::{KolmeRuntimeCommitRequest, KolmeRuntimeCommitSignedBr
 pub use runtime_pipeline::{
     RuntimeCommitFinalityProjection, RuntimeCommitLifecycleRecord, RuntimeCommitPipeline,
 };
-
-/// Abstract client interface for Kolme runtime commit submission.
-pub trait KolmeRuntimeCommitClient {
-    /// Submits one deterministic runtime commit request.
-    fn submit_commit(
-        &mut self,
-        request: &KolmeRuntimeCommitRequest,
-    ) -> Result<KolmeRuntimeCommitOutcome, KolmeRuntimeCommitError>;
-}
-
-/// Transport connection abstraction for consuming notifications text messages.
-pub trait KolmeRuntimeCommitNotificationsConnection {
-    /// Reads the next notifications text message.
-    ///
-    /// Returns `Ok(None)` when the current websocket connection is closed.
-    fn read_text_message(&mut self) -> Result<Option<String>, KolmeRuntimeCommitProviderError>;
-}
-
-/// Connector abstraction for establishing notifications websocket connections.
-pub trait KolmeRuntimeCommitNotificationsConnector {
-    /// Concrete connection type returned by the connector.
-    type Connection: KolmeRuntimeCommitNotificationsConnection;
-
-    /// Connects to one websocket notifications URL.
-    fn connect(
-        &mut self,
-        notifications_url: &str,
-    ) -> Result<Self::Connection, KolmeRuntimeCommitProviderError>;
-}
-
-/// Provider interface consumed by the adapter-backed runtime commit client.
-pub trait KolmeRuntimeCommitProvider {
-    /// Submits canonical wire payload with deterministic idempotency key.
-    fn submit_runtime_commit(
-        &mut self,
-        wire_payload: &str,
-        idempotency_key: &str,
-    ) -> Result<KolmeRuntimeCommitProviderOutcome, KolmeRuntimeCommitProviderError>;
-}
-
-/// Transport abstraction used by the live provider bridge to reach Kolme backends.
-pub trait KolmeRuntimeCommitProviderTransport {
-    /// Submits one runtime commit payload to the configured provider endpoint.
-    fn submit_runtime_commit(
-        &mut self,
-        base_url: &str,
-        submit_path: &str,
-        wire_payload: &str,
-        idempotency_key: &str,
-    ) -> Result<String, KolmeRuntimeCommitProviderError>;
-}
-
-/// Transport abstraction for querying runtime commit finality from a live backend.
-pub trait KolmeRuntimeCommitFinalityTransport {
-    /// Fetches one finality response payload for the provided commit identifier.
-    fn fetch_runtime_commit_finality(
-        &mut self,
-        base_url: &str,
-        status_path: &str,
-        commit_id: &str,
-    ) -> Result<String, KolmeRuntimeCommitProviderError>;
-}
-
-/// Transport abstraction for querying `/block/{height}` fallback responses.
-pub trait KolmeRuntimeCommitBlockFallbackTransport {
-    /// Fetches one block response payload for the provided height.
-    fn fetch_block_by_height(
-        &mut self,
-        base_url: &str,
-        block_path_template: &str,
-        height: u64,
-    ) -> Result<String, KolmeRuntimeCommitProviderError>;
-}
 
 #[cfg(test)]
 mod tests {
