+++ crates/kamn-core/tests/kolme_runtime_commit_extraction_boundary.rs
@@ -22,6 +22,7 @@ const RUNTIME_NOTIFICATIONS_WS_SRC: &str =
 const RUNTIME_PIPELINE_SRC: &str = include_str!("../src/kolme_runtime_commit/runtime_pipeline.rs");
 const RUNTIME_ERRORS_SRC: &str = include_str!("../src/kolme_runtime_commit/errors.rs");
 const RUNTIME_OUTCOMES_SRC: &str = include_str!("../src/kolme_runtime_commit/outcomes.rs");
+const RUNTIME_INTERFACES_SRC: &str = include_str!("../src/kolme_runtime_commit/interfaces.rs");
 
 #[test]
 fn unit_runtime_commit_extraction_boundary_removes_local_finality_glue_wrappers() {
@@ -106,6 +107,14 @@ fn unit_runtime_commit_extraction_boundary_removes_local_finality_glue_wrappers(
     assert!(!RUNTIME_COMMIT_SRC.contains(
         "impl From<KamnKolmeRuntimeProviderOutcome> for KolmeRuntimeCommitProviderOutcome"
     ));
+    // Regression: #1964
+    assert!(!RUNTIME_COMMIT_SRC.contains("pub trait KolmeRuntimeCommitClient"));
+    assert!(!RUNTIME_COMMIT_SRC.contains("pub trait KolmeRuntimeCommitNotificationsConnection"));
+    assert!(!RUNTIME_COMMIT_SRC.contains("pub trait KolmeRuntimeCommitNotificationsConnector"));
+    assert!(!RUNTIME_COMMIT_SRC.contains("pub trait KolmeRuntimeCommitProvider"));
+    assert!(!RUNTIME_COMMIT_SRC.contains("pub trait KolmeRuntimeCommitProviderTransport"));
+    assert!(!RUNTIME_COMMIT_SRC.contains("pub trait KolmeRuntimeCommitFinalityTransport"));
+    assert!(!RUNTIME_COMMIT_SRC.contains("pub trait KolmeRuntimeCommitBlockFallbackTransport"));
     assert!(!RUNTIME_COMMIT_SRC.contains("impl RuntimeCommitPipeline"));
     assert!(!RUNTIME_COMMIT_SRC.contains("impl InMemoryKolmeRuntimeCommitClient"));
     assert!(!RUNTIME_COMMIT_SRC
@@ -323,6 +332,7 @@ fn regression_runtime_commit_extraction_boundary_keeps_direct_helper_delegation(
     assert!(RUNTIME_COMMIT_SRC.contains("mod notifications_websocket;"));
     assert!(RUNTIME_COMMIT_SRC.contains("mod errors;"));
     assert!(RUNTIME_COMMIT_SRC.contains("mod outcomes;"));
+    assert!(RUNTIME_COMMIT_SRC.contains("mod interfaces;"));
     assert!(RUNTIME_COMMIT_SRC.contains("pub use request_envelope::{"));
     assert!(RUNTIME_COMMIT_SRC.contains("pub use runtime_pipeline::{"));
     assert!(
@@ -345,6 +355,7 @@ fn regression_runtime_commit_extraction_boundary_keeps_direct_helper_delegation(
     assert!(RUNTIME_COMMIT_SRC.contains("pub use notifications_websocket::{"));
     assert!(RUNTIME_COMMIT_SRC.contains("pub use errors::{"));
     assert!(RUNTIME_COMMIT_SRC.contains("pub use outcomes::{"));
+    assert!(RUNTIME_COMMIT_SRC.contains("pub use interfaces::{"));
     assert!(RUNTIME_ERRORS_SRC.contains("pub enum KolmeRuntimeCommitTransportErrorKind"));
     assert!(RUNTIME_ERRORS_SRC.contains("pub enum KolmeRuntimeCommitProviderError"));
     assert!(RUNTIME_ERRORS_SRC.contains("pub enum KolmeRuntimeCommitError"));
@@ -353,4 +364,11 @@ fn regression_runtime_commit_extraction_boundary_keeps_direct_helper_delegation(
     assert!(RUNTIME_OUTCOMES_SRC.contains("pub struct KolmeRuntimeCommitProviderReceipt"));
     assert!(RUNTIME_OUTCOMES_SRC.contains("pub enum KolmeRuntimeCommitProviderOutcome"));
     assert!(RUNTIME_OUTCOMES_SRC.contains("pub enum KolmeRuntimeCommitNotificationEvent"));
+    assert!(RUNTIME_INTERFACES_SRC.contains("pub trait KolmeRuntimeCommitClient"));
+    assert!(RUNTIME_INTERFACES_SRC.contains("pub trait KolmeRuntimeCommitNotificationsConnection"));
+    assert!(RUNTIME_INTERFACES_SRC.contains("pub trait KolmeRuntimeCommitNotificationsConnector"));
+    assert!(RUNTIME_INTERFACES_SRC.contains("pub trait KolmeRuntimeCommitProvider"));
+    assert!(RUNTIME_INTERFACES_SRC.contains("pub trait KolmeRuntimeCommitProviderTransport"));
+    assert!(RUNTIME_INTERFACES_SRC.contains("pub trait KolmeRuntimeCommitFinalityTransport"));
+    assert!(RUNTIME_INTERFACES_SRC.contains("pub trait KolmeRuntimeCommitBlockFallbackTransport"));
 }
